refactor: rename moment extend file
